implement new terminal components which use the container / deployment / stack specific permissioned endpoints
